Fix one farming circular ref
